Extract video thumbnail data wrapping into a shared Attachment extension
The VideoThumbnailImageData construction with the thumbnail-or-asset guard was
duplicated across both UI kits (message list, media grids, file/quoted previews).
Unify it in Attachment.videoThumbnailImageData and move the data class plus the
extension into their own file; call sites keep their own CDN resizing and null
fallback.
