+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/extensions/internal/AttachmentExtensions.kt
@@ -21,7 +21,7 @@ import io.getstream.chat.android.client.utils.attachment.isImage
 import io.getstream.chat.android.client.utils.attachment.isVideo
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.models.Attachment
-import io.getstream.chat.android.ui.common.images.internal.VideoThumbnailImageData
+import io.getstream.chat.android.ui.common.images.internal.videoThumbnailImageData
 import io.getstream.chat.android.ui.common.images.resizing.applyStreamCdnImageResizingIfEnabled
 
 /**
@@ -33,7 +33,7 @@ import io.getstream.chat.android.ui.common.images.resizing.applyStreamCdnImageRe
  * For image attachments, [Attachment.imageUrl] is used.
  * For video attachments when thumbnails are enabled, the server [Attachment.thumbUrl] is used,
  * falling back to a frame extracted from the [Attachment.assetUrl] video when the thumbnail is
- * missing or fails to load (see [VideoThumbnailImageData]).
+ * missing or fails to load (see [videoThumbnailImageData]).
  */
 @get:Composable
 internal val Attachment.imagePreviewData: Any?
@@ -44,11 +44,7 @@ internal val Attachment.imagePreviewData: Any?
                 ?: upload
         isVideo() && ChatTheme.videoThumbnailsEnabled -> {
             val thumbnailUrl = thumbUrl?.applyStreamCdnImageResizingIfEnabled(ChatTheme.streamCdnImageResizing)
-            if (thumbnailUrl != null || assetUrl != null) {
-                VideoThumbnailImageData(thumbnailUrl = thumbnailUrl, videoUrl = assetUrl)
-            } else {
-                upload
-            }
+            videoThumbnailImageData(thumbnailUrl) ?: upload
         }
         else -> null
     }
